pyproject.toml
[tool.poetry]
name = "barcode-check-scf-duplicates-azure-func"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
azure-functions = "^1.21.3"
requests = "^2.32.3"
jinja2 = "^3.1.5"
python-dotenv = "^1.0.1"
azure-identity = "^1.19.0"
azure-keyvault-secrets = "^4.9.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
sqlalchemy = "^2.0.38"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.4"
mypy = "^1.15.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
